+++ src/Morphology/Morphology.jl
@@ -1,18 +1,22 @@
 module Morphology
 
-"""
-    dummy_morphology_function()
-
-Example function for morphology module
-
-!!! todo "Delete this once real functions are added"
-    This function should be removed when real morphology functions are moved into this module.
+import ImageMorphology: ImageMorphology
 
 """
-function dummy_morphology_function()
-    return "This is a dummy morphology function."
+    imregionalmin(img, conn=2)
+
+Compute the regional minima of the image `img` using the connectivity `conn`.
+
+Returns a bitmatrix of the same size as `img` with the regional minima.
+
+# Arguments
+- `img`: Image object
+- `conn`: Neighborhood connectivity; in 2D, 1 = 4-neighborhood and 2 = 8-neighborhood
+"""
+function imregionalmin(img, conn=2)
+    return ImageMorphology.local_minima(img; connectivity=conn) .> 0
 end
 
-export dummy_morphology_function
+export imregionalmin
 
 end
