+++ src/utils.jl
@@ -85,20 +85,6 @@ function impose_minima(
     return 1 .- sk_morphology.reconstruction(1 .- marker, 1 .- mask)
 end
 
-"""
-    imregionalmin(img, conn=2)
-
-Compute the regional minima of the image `img` using the connectivity `conn`.
-
-Returns a bitmatrix of the same size as `img` with the regional minima.
-
-# Arguments
-- `img`: Image object
-- `conn`: Neighborhood connectivity; in 2D, 1 = 4-neighborhood and 2 = 8-neighborhood
-"""
-function imregionalmin(img, conn=2)
-    return ImageMorphology.local_minima(img; connectivity=conn) .> 0
-end
 
 """
     bwdist(bwimg)
